+++ examples/matrixmultiplication3/src/at/illecker/rootbeer/examples/matrixmultiplication3/MatrixMultiplication3Kernel.java
@@ -258,7 +258,7 @@ public class MatrixMultiplication3Kernel implements Kernel {
     Context context = rootbeer.createDefaultContext();
     Stopwatch watch = new Stopwatch();
     watch.start();
-    rootbeer.run(kernel, new ThreadConfig(blockSize, gridSize, (long) blockSize
+    rootbeer.run(kernel, new ThreadConfig(blockSize, gridSize, blockSize
         * gridSize), context);
     watch.stop();
